core+ai+pipeline+updater: build cleanly under Apple Clang 21 + -Werror
Apple's clang shipped in Xcode 26 (clang 21) is stricter than the
Linux/gcc-13 toolchain CI compiles against — both on warning
classes already enforced and on a handful of structural rules that
trigger real errors. The whole tree now builds with the existing
`-Wall -Wextra -Wpedantic -Wshadow -Wconversion -Wsign-conversion
-Wformat=2 -Wimplicit-fallthrough -Werror` flag set under
AppleClang 21 + Apple linker, so contributors developing on macOS
can run the project's own warning gate locally without flipping
escape hatches.

90 errors → 0. No `-Wno-*` suppressions added; every fix is at
the source. No frozen-header (include/souxmar-c/*) touches.

Per-category summary:

  47 × -Wmissing-field-initializers
    `ToolError` gained a `suggestion` field after the call sites
    were written. Default-initialise `suggestion = ""` once in
    include/souxmar/ai/tool.h so the existing brace-init at every
    ai/tools/*.cpp site stops triggering the warning. Zero call-
    site edits needed.

  16 × -Wsign-conversion
    Loop indices declared `int` while indexing into containers
    with `size_t` interfaces (`std::array`, `std::vector`,
    `std::span`). Changed the loop variable type to `std::size_t`
    in c_abi_geometry.cpp, c_abi_surface_stream.cpp,
    surface_stream.cpp, mesh_quality.cpp (the edges_with_opposites
    table), and registry_dispatcher.cpp's length-decode loop.

   1 × -Wswitch
    list_plugins.cpp's kind_token switch missed `CapabilityKind::
    Reader` — added since Sprint 6 push 4. The fallthrough return
    "unknown" was masking it; now exhaustive.

   1 × -Wunused-function
    apply_pipeline_diff.cpp had a stub `stage_id_of` whose own
    comment said it returned nullptr unconditionally; its real
    replacement `stage_id_view` was right below. Deleted.

  18 × undeclared identifier
    registry_dispatcher.cpp called `souxmar_mesh_free`,
    `souxmar_field_free`, `souxmar_geometry_free` without including
    their declaring headers. Older clang was tolerating the
    transitive include path through `souxmar-c/mesher.h`. Added
    explicit `#include "souxmar-c/{mesh,field,geometry}.h"` lines.

   1 × undeclared identifier
    `fmt::join` moved to <fmt/ranges.h> in fmt v9. Added the
    include in dag.cpp.

   1 × ambiguous lookup
    `Mesh::add_cell` called unqualified `num_nodes(type)`. Newer
    clang resolves the unqualified name to the member function
    `Mesh::num_nodes()` (taking no args) first and errors on the
    extra argument before ADL gets a chance at the free function.
    Qualified the call as `::souxmar::core::num_nodes(type)`.

   2 × deleted defaulted move
    `Registry` holds a `mutable std::shared_mutex mu_`, which is
    non-movable. `Registry::Registry(Registry&&) = default` (and
    the move-assign counterpart) is implicitly deleted by the
    compiler; clang 21 errors on `= default` for an
    implicitly-deleted special function. Replaced with hand-rolled
    moves that transfer the entries map + current-plugin scratch
    fields and leave each instance's mutex default-constructed.
    Caller is responsible for serialising concurrent use across a
    move (Registry is one-owner in practice — grep confirms no
    actual move uses anywhere).

   5 × private member access from extern "C"
    The C ABI trampolines `souxmar_registry_add_{mesher,solver,
    writer,postproc,reader}` reach into `Registry::current_plugin_
    id_` (a private member) to credit the active plugin. Newer
    clang correctly rejects this — the trampolines are free
    functions, not members or friends. Added a public
    `current_plugin_id() const noexcept` accessor (returning
    string_view) and routed the trampolines through it. This is
    the access pattern the header comment already described:
    "Called by the extern wrappers — kept here so the wrapper can
    avoid friend tricks."

Verified end-to-end: clean rebuild produces all 7 static libs
(libsouxmar-{core,plugin,pipeline,ai,crypto,update,c-bridge}.a),
106-task Ninja graph completes with zero warnings.

+++ src/updater/manifest.cpp
@@ -420,8 +420,11 @@ std::vector<ManifestValidationIssue> validate_manifest(const Manifest& m) {
   std::unordered_set<std::uint16_t> seen_pairs;
   for (std::size_t i = 0; i < m.artifacts.size(); ++i) {
     const auto& a = m.artifacts[i];
-    const std::uint16_t pair =
-        (static_cast<std::uint16_t>(a.os) << 8) | static_cast<std::uint16_t>(a.arch);
+    // (os << 8) | arch integer-promotes through int; cast back to
+    // uint16_t to satisfy -Wconversion. Both enums are uint8_t-backed
+    // (manifest.h), so the packed value always fits in 16 bits.
+    const std::uint16_t pair = static_cast<std::uint16_t>((static_cast<std::uint16_t>(a.os) << 8)
+                                                          | static_cast<std::uint16_t>(a.arch));
     if (!seen_pairs.insert(pair).second) {
       out.push_back({ManifestIssueSeverity::Error,
                      artifact_field(i, "os+arch"),
